Changed tests to write into separate tfrecords directory. InterpDataSet now handles validation_size loading internally.

+++ data/interp/interp_data_test.py
@@ -18,6 +18,7 @@ class TestInterpDataSet(unittest.TestCase):
     def setUp(self):
         cur_dir = os.path.dirname(__file__)
         self.data_directory = os.path.join(cur_dir, 'davis', 'test_data')
+        self.tf_record_directory = os.path.join(self.data_directory, 'tfrecords')
 
         # Test paths.
         self.expected_image_paths_0 = [
@@ -41,7 +42,7 @@ class TestInterpDataSet(unittest.TestCase):
         self.sess = tf.Session(config=config)
 
     def test_maximum_shot_len(self):
-        data_set = DavisDataSet(self.data_directory, [[1]], maximum_shot_len=3)
+        data_set = DavisDataSet(self.tf_record_directory, [[1]], maximum_shot_len=3)
         image_paths = [
             ['a0', 'a1', 'a2'],
             ['b0', 'b1', 'b2', 'b3', 'b4', 'b5', 'b6'],
@@ -61,7 +62,7 @@ class TestInterpDataSet(unittest.TestCase):
     def test_val_split(self):
 
         # Sequences of lengths 3, 4, 5.
-        data_set = DavisDataSet(self.data_directory, [[1], [1, 0], [1, 0, 0]])
+        data_set = DavisDataSet(self.tf_record_directory, [[1], [1, 0], [1, 0, 0]])
         image_paths = [
             ['a0', 'a1', 'a2'],
             ['b0', 'b1', 'b2', 'b3', 'b4', 'b5', 'b6'],
@@ -82,7 +83,7 @@ class TestInterpDataSet(unittest.TestCase):
     def test_val_split_all(self):
 
         # Sequences of lengths 3, 4, 5.
-        data_set = DavisDataSet(self.data_directory, [[1], [1, 0], [1, 0, 0]])
+        data_set = DavisDataSet(self.tf_record_directory, [[1], [1, 0], [1, 0, 0]])
         image_paths = [
             ['a0', 'a1', 'a2'],
             ['b0', 'b1', 'b2', 'b3', 'b4', 'b5', 'b6'],
@@ -99,7 +100,7 @@ class TestInterpDataSet(unittest.TestCase):
         self.assertListEqual(train, expected_train)
 
     def test_data_read_write(self):
-        data_set = DavisDataSet(self.data_directory, [[1]], batch_size=2)
+        data_set = DavisDataSet(self.tf_record_directory, [[1]], batch_size=2)
         data_set.preprocess_raw(self.data_directory, shard_size=1)
 
         output_paths = data_set.get_tf_record_names()
@@ -123,7 +124,7 @@ class TestInterpDataSet(unittest.TestCase):
             self.assertTupleEqual(np.shape(next_sequence), (2, 3, 264, 470, 3))
 
     def test_val_data_read_write(self):
-        data_set = DavisDataSet(self.data_directory, [[1]], batch_size=2)
+        data_set = DavisDataSet(self.tf_record_directory, [[1]], batch_size=2)
         data_set.preprocess_raw(self.data_directory, shard_size=5, validation_size=2)
 
         output_paths = data_set.get_tf_record_names()
@@ -157,7 +158,7 @@ class TestInterpDataSet(unittest.TestCase):
         """
         Tests for the case where multiple inbetween_location configs are provided.
         """
-        data_set = DavisDataSet(self.data_directory, [[1], [1, 0, 0]], batch_size=1)
+        data_set = DavisDataSet(self.tf_record_directory, [[1], [1, 0, 0]], batch_size=1)
         data_set.preprocess_raw(self.data_directory, shard_size=1)
 
         output_paths = data_set.get_tf_record_names()
@@ -200,14 +201,19 @@ class TestInterpDataSet(unittest.TestCase):
         self.assertEqual(num_sparse_sequences, 4)
 
     def tearDown(self):
-        data_set = DavisDataSet(self.data_directory, [[1]], batch_size=2)
+        data_set = DavisDataSet(self.tf_record_directory, [[1]], batch_size=2)
         output_paths = data_set.get_tf_record_names()
         for output_path in output_paths:
             if os.path.isfile(output_path):
                 os.remove(output_path)
 
-        if os.listdir(data_set.get_tf_record_dir()) == []:
-            shutil.rmtree(data_set.get_tf_record_dir())
+        json_path = os.path.join(self.tf_record_directory, 'val_split.json')
+        if os.path.isfile(json_path):
+            os.remove(json_path)
+
+        if os.path.isdir(data_set.get_tf_record_dir()):
+            if os.listdir(data_set.get_tf_record_dir()) == []:
+                shutil.rmtree(data_set.get_tf_record_dir())
 
 
 if __name__ == '__main__':
